Make converter separate from BaseModel
Rather than inheriting BaseModel and forcing usage to use XmlBaseModel,
the main class only stores XmlAttribute information and exposes
conversion methods.

Now a converter called PydanticXmlConverter is provided that parses
to/from XML to regular Pydantic classes.

For attributes to be persisted, the same converter must be used for
parsing/generating, or the XmlAttribute list must be carried over.